Sealing CheckingAccount and SavingsAccount.
Abstracting BankAccount.

Adding TestBankAccount to hack a way through to BankAccount objects for tests that don't specifically address narrower implementations of BankAccount.

Fixed BankAccount initialization in test classes to use TestBankAccount.

Updated the form to select from concrete BankAccount derivations. This is purposefully done in a hard-to-maintain way as part of the demonstration.